Fix ordering to prefer latest package set for legacy manifests
When determining bounds for a legacy package set import, we want to
look at the latest package set that included it to get the most
recent bounds that would work for it. This is added to the bounds
from the published packages.dhall and/or bower.json in the repo.