Overhaul and modernize Nix flake setup (#706)
* add envrc for direnv users

this was already used by several developers, this just makes it
official

* replace run-tests-script with nix flake check

this also allows us to remove our single 'spago test' invocation
in a github workflow, so all CI checks are in the Nix checks for
Garnix to execute

while i was at it, i removed the nix flakes wrapper, shell.nix,
and flake-compat inclusions as at this point pretty much everyone
is on a flake-compatible version of Nix

* extract wiremock mappings

* migrate to buildNpmPackage from slimlock

slimlock was the homegrown npm dependency builder from ps-overlay
from before when buildNpmPackage was a suitable choice for the
registry, but at this point it's well-suited to be the standard
choice for us.

* overhaul nix flake setup

our homegrown nix setup has diverged quite a bit from modern
best practices and is hard to read. it also included some
unnecessary helpers (like run-vm to work on a local vm) which
we can remove.

this change overhauls the full setup while maintaining the
existing tests to make it simpler, more performant, and more
usable for contributors.

the deployments are still simple ('colmena apply').